Move the compounding calculator into its own Django application.
