+++ siteconfig/urls.py
@@ -18,4 +18,5 @@ from django.conf.urls import include, url
 
 urlpatterns = [
   url(r'^', include('charting.urls')),
+  url(r'^', include('compound.urls')),
 ]
